wakeupcalls/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
twilio==8.10.3
requests==2.31.0
python-decouple==3.8
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
django-extensions==3.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
boto3==1.34.0
watchtower==3.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
whitenoise==6.6.0
dj-database-url==2.1.0
